requirements.txt
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
futures==3.1.1
google-api-core==1.5.1
google-auth==1.5.1
google-cloud-vision==0.34.0
googleapis-common-protos==1.5.5
grpcio==1.16.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.3
opencv-python==3.4.3.18
Pillow==5.3.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pybase64==0.4.0
pytz==2018.7
requests==2.20.0
rsa==4.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
